Inkluderer loader i felles-komponent.
