+++ src/components/_common/product-panel/ProductPanelExpandable.module.scss
@@ -1,15 +1,5 @@
 @use 'src/common' as common;
 
-.detailsLoader {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    :global(.navds-loader) {
-        margin-bottom: 1rem;
-    }
-}
-
 .accordionItem {
     border-bottom: 1px common.$a-border-divider solid;
 
@@ -66,4 +56,14 @@
 
 .copyLink {
     margin-bottom: 1rem;
+}
+
+.loader {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    :global(.navds-loader) {
+        margin-bottom: 1rem;
+    }
 }
